# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fleet Control API
  version: 0.2.1
  description: |
    Public web API consumed by the Fleet web UI. The API aggregates multiple
    device roles (audio players, video/CEC, Zigbee, cameras) and exposes a
    consistent contract for UI integrations. Schemas align with the live device
    capabilities matrix to ensure client compatibility.
  contact:
    name: Fleet Platform Team
    email: platform@fleet.example
servers:
  - url: /api
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  headers:
    CorrelationId:
      description: Correlation identifier for tracing requests end-to-end.
      schema:
        type: string
    RateLimitLimit:
      description: Maximum number of requests allowed in the current window.
      schema:
        type: integer
        format: int32
    RateLimitRemaining:
      description: Remaining requests in the current window.
      schema:
        type: integer
        format: int32
    RetryAfterSeconds:
      description: Seconds until clients may retry after a 429 response.
      schema:
        type: integer
        format: int32
...
  responses:
    UnauthorizedError:
      description: Authentication failed or credentials missing.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: UNAUTHENTICATED
                message: Authentication token missing or invalid.
                correlationId: 00000000-0000-0000-0000-000000000000
    ForbiddenError:
      description: Authenticated user does not have permission to access the resource.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Requested resource does not exist.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: One or more request parameters failed validation.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid:
              value:
                code: INVALID_REQUEST
                message: Provided payload failed validation.
                correlationId: 11111111-1111-1111-1111-111111111111
    RateLimitError:
      description: Request rate limit exceeded.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
        x-ratelimit-limit:
          $ref: '#/components/headers/RateLimitLimit'
        x-ratelimit-remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        retry-after:
          $ref: '#/components/headers/RetryAfterSeconds'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Unexpected server error occurred.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - correlationId
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable summary of the error.
        hint:
          type: string
          nullable: true
          description: Optional client-facing recommendation.
        correlationId:
          type: string
          format: uuid
          description: Correlates the request across systems.
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Additional contextual information for debugging.
...
    ProxySettings:
      type: object
      required:
        - upstreamUrl
        - authMode
        - heartbeatIntervalSeconds
      properties:
        upstreamUrl:
          type: string
          format: uri
        authMode:
          type: string
          enum:
            - none
            - basic
            - token
        heartbeatIntervalSeconds:
          type: integer
          format: int32
          minimum: 5
    ProxySettingsPatch:
      type: object
      properties:
        upstreamUrl:
          type: string
          format: uri
        authMode:
          type: string
          enum:
            - none
            - basic
            - token
        heartbeatIntervalSeconds:
          type: integer
          format: int32
          minimum: 5
          maximum: 3600
    SettingsPairingState:
      type: object
      required:
        - active
        - expiresAt
        - ticketId
        - candidates
      properties:
        active:
          type: boolean
        expiresAt:
          type: string
          format: date-time
          nullable: true
        ticketId:
          type: string
          nullable: true
        candidates:
          type: array
          items:
            type: object
            required:
              - id
              - model
              - signal
            properties:
              id:
                type: string
              model:
                type: string
              signal:
                type: integer
                format: int32
    OperatorAccount:
      type: object
      required:
        - id
        - email
        - roles
        - invitedAt
        - status
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
        invitedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum:
            - active
            - pending
            - removed
    SettingsState:
      type: object
      required:
        - proxy
        - allowedOrigins
        - pairing
        - operators
        - apiTokenPreview
        - updatedAt
      properties:
        proxy:
          $ref: '#/components/schemas/ProxySettings'
        allowedOrigins:
          type: array
          items:
            type: string
        pairing:
          $ref: '#/components/schemas/SettingsPairingState'
        operators:
          type: array
          items:
            $ref: '#/components/schemas/OperatorAccount'
        apiTokenPreview:
          type: string
        updatedAt:
          type: string
          format: date-time
          nullable: true
    ProxyUpdateRequest:
      type: object
      required:
        - proxy
      properties:
        proxy:
          $ref: '#/components/schemas/ProxySettingsPatch'
    AllowedOriginsRequest:
      type: object
      required:
        - origins
      properties:
        origins:
          type: array
          items:
            type: string
    PairingStartRequest:
      type: object
      required:
        - networkRole
        - expiresInSeconds
      properties:
        networkRole:
          type: string
          enum:
            - audio
            - video
            - lighting
            - sensor
        expiresInSeconds:
          type: integer
          format: int32
          minimum: 60
          maximum: 3600
    PairingClaimRequest:
      type: object
      properties:
        alias:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties:
            type: string
    InviteOperatorRequest:
      type: object
      required:
        - email
        - roles
      properties:
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
...
paths:
...
  /settings:
    get:
      tags:
        - Settings
      summary: Retrieve operator settings, proxy configuration, and pairing status.
      description: Retrieve operator settings, proxy configuration, and pairing status.
      operationId: getSettings
      responses:
        '200':
          description: Settings payload for the control plane.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsState'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/proxy:
    patch:
      tags:
        - Settings
      summary: Update proxy configuration settings.
      description: Update proxy configuration settings.
      operationId: updateProxySettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyUpdateRequest'
      responses:
        '202':
          description: Proxy update accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - proxy
                  - updatedAt
                properties:
                  proxy:
                    $ref: '#/components/schemas/ProxySettings'
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/api-token:
    post:
      tags:
        - Settings
      summary: Rotate the API bearer token used by operator integrations.
      description: Rotate the API bearer token used by operator integrations.
      operationId: rotateApiToken
      responses:
        '202':
          description: Token rotation accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - tokenPreview
                  - rotatedAt
                properties:
                  status:
                    type: string
                  tokenPreview:
                    type: string
                  rotatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/allowed-origins:
    put:
      tags:
        - Settings
      summary: Replace the list of allowed CORS origins for the API.
      description: Replace the list of allowed CORS origins for the API.
      operationId: updateAllowedOrigins
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllowedOriginsRequest'
      responses:
        '202':
          description: Allowed origins updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - allowedOrigins
                  - updatedAt
                properties:
                  allowedOrigins:
                    type: array
                    items:
                      type: string
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/pairing/start:
    post:
      tags:
        - Settings
      summary: Start operator pairing flow for onboarding new devices.
      description: Start operator pairing flow for onboarding new devices.
      operationId: startSettingsPairing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairingStartRequest'
      responses:
        '202':
          description: Pairing flow started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsPairingState'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/pairing/cancel:
    post:
      tags:
        - Settings
      summary: Cancel an in-progress pairing flow.
      description: Cancel an in-progress pairing flow.
      operationId: cancelSettingsPairing
      responses:
        '202':
          description: Pairing flow cancelled.
          content:
            application/json:
              schema:
                type: object
                required:
                  - cancelled
                  - updatedAt
                properties:
                  cancelled:
                    type: boolean
                  updatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/pairing/{candidateId}/claim:
    post:
      tags:
        - Settings
      summary: Claim a discovered device and record pairing outcome.
      description: Claim a discovered device and record pairing outcome.
      operationId: claimPairingCandidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairingClaimRequest'
      responses:
        '202':
          description: Pairing candidate claim accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - accepted
                  - candidateId
                  - metadata
                  - updatedAt
                properties:
                  accepted:
                    type: boolean
                  candidateId:
                    type: string
                  metadata:
                    type: object
                    additionalProperties:
                      type: string
                  updatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/operators:
    post:
      tags:
        - Settings
      summary: Invite a new operator account with specified roles.
      description: Invite a new operator account with specified roles.
      operationId: inviteOperator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteOperatorRequest'
      responses:
        '201':
          description: Operator invited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorAccount'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Operator already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/operators/{operatorId}:
    delete:
      tags:
        - Settings
      summary: Remove an operator account.
      description: Remove an operator account.
      operationId: removeOperator
      parameters:
        - name: operatorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Operator removal accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - removed
                  - operatorId
                  - updatedAt
                properties:
                  removed:
                    type: boolean
                  operatorId:
                    type: string
                  updatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'